personal workspace: make folder tree optional
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@28390 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/Object/classes/class.ilObject2GUI.php
@@ -408,7 +408,10 @@ abstract class ilObject2GUI extends ilObjectGUI
 
 				// object
 				$object = ilObjectFactory::getInstanceByObjId($node["obj_id"], false);
-			    $object->delete();
+				if($object)
+				{
+					$object->delete();
+				}
 			}
 
 			$this->ctrl->setParameter($this, "wsp_id", $parent);
